8th task: defining query service interface as C"Q"RS
